Use Python GIL
