--- algo/F2,rj,pmtn,Cmax/file_pa.txt
@@ -1,10 +1,0 @@
-9 1 2
-1 3 7
-6 1 1
-6 2 10
-10 7 9
-4 3 6
-2 5 8
-1 3 8
-6 3 8
-5 3 4
